* changes:
  [ALPS08634904] r47p0: fix feature conflict
  [ALPS08634904] [c6260] Mali Avalon DDK r47p0-01eac0

MTK-Commit-Id: 182c435196df8627a9193b3c32f7294d12c92103

+++ gpu_mali/mali_avalon/mali-r47p0/drivers/gpu/arm/midgard/backend/gpu/mali_kbase_js_backend.c
@@ -171,15 +171,15 @@ static enum hrtimer_restart timer_callback(struct hrtimer *timer)
 					 * now. Hard stop the slot.
 					 */
 #if !KBASE_DISABLE_SCHEDULING_HARD_STOPS
-					int ms = js_devdata->scheduling_period_ns / 1000000u;
+					u32 ms = js_devdata->scheduling_period_ns / 1000000u;
 #if IS_ENABLED(CONFIG_MALI_MTK_DEBUG_DUMP)
 					mtk_common_debug(MTK_COMMON_DBG_DUMP_PM_STATUS, -1, MTK_DBG_HOOK_NA);
 					mtk_common_debug(MTK_COMMON_DBG_DUMP_INFRA_STATUS, -1, MTK_DBG_HOOK_NA);
 #endif /* CONFIG_MALI_MTK_DEBUG_DUMP */
 					dev_warn(
 						kbdev->dev,
-						"JS: Job Hard-Stopped (took more than %lu ticks at %lu ms/tick)",
-						(unsigned long)ticks, (unsigned long)ms);
+						"JS: Job Hard-Stopped (took more than %u ticks at %u ms/tick)",
+						ticks, ms);
 					kbase_job_slot_hardstop(atom->kctx, s, atom);
 #endif
 				} else if (ticks == gpu_reset_ticks) {
@@ -210,15 +210,15 @@ static enum hrtimer_restart timer_callback(struct hrtimer *timer)
 					 * ticks. Hard stop the slot.
 					 */
 #if !KBASE_DISABLE_SCHEDULING_HARD_STOPS
-					int ms = js_devdata->scheduling_period_ns / 1000000u;
+					u32 ms = js_devdata->scheduling_period_ns / 1000000u;
 #if IS_ENABLED(CONFIG_MALI_MTK_DEBUG_DUMP)
 					mtk_common_debug(MTK_COMMON_DBG_DUMP_PM_STATUS, -1, MTK_DBG_HOOK_NA);
 					mtk_common_debug(MTK_COMMON_DBG_DUMP_INFRA_STATUS, -1, MTK_DBG_HOOK_NA);
 #endif /* CONFIG_MALI_MTK_DEBUG_DUMP */
 					dev_warn(
 						kbdev->dev,
-						"JS: Job Hard-Stopped (took more than %lu ticks at %lu ms/tick)",
-						(unsigned long)ticks, (unsigned long)ms);
+						"JS: Job Hard-Stopped (took more than %u ticks at %u ms/tick)",
+						ticks, ms);
 					kbase_job_slot_hardstop(atom->kctx, s, atom);
 #endif
 				} else if (ticks == js_devdata->gpu_reset_ticks_dumping) {
